Validando ProjectMembers

+++ app/Http/Controllers/ProjectController.php
@@ -31,7 +31,7 @@ class ProjectController extends Controller
     }
     public function index()
     {
-    	 return $this->repository->all();
+    	 return $this->repository->findWhere(['owner_id'=>\Authorizer::getResourceOwnerId()]);
 	}
     /**
      * @param Request $request
@@ -47,6 +47,10 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if ($this->checkProjectOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidem'];
+        }
         return $this->repository->find($id);
     }
     /**
@@ -56,6 +60,10 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if ($this->checkProjectOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidem'];
+        }
         return $this->service->update($request->all(), $id);
     }
     /**
@@ -64,7 +72,17 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if ($this->checkProjectOwner($id) == false)
+        {
+            return ['error' => 'Access Forbidem'];
+        }
         return $this->repository->delete($id);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+        return $this->repository->isOwner($projectId, $userId);
     }
 }
